register.jsp

+++ pw_practica2/src/main/java/model/DAOS/EspectaculoDAO.java
@@ -2,7 +2,7 @@ package model.DAOS;
 
 import model.Data.Espectaculo;
 
-public interface EspectaculoDAO extends DAO<Espectaculo,int>{
+public interface EspectaculoDAO extends DAO<Espectaculo,Long>{
 
 	
 }
