+++ pw_practica2/src/main/java/model/ficheros/DMEspectaculoDAO.java
@@ -68,7 +68,7 @@ ArrayList<Espectaculo> espectaculos = new ArrayList<Espectaculo>();
 	}
 
 	@Override
-	public Espectaculo obtener(int id) {
+	public Espectaculo obtener(Long id) {
 		
 		for(Espectaculo UsuarioAuxiliar : espectaculos) {
 			if(UsuarioAuxiliar.getIdEspectaculo() == id) {
